@@ -25,8 +25,7 @@
     "url": "https://github.com/springload/draftail/issues"
   },
   "homepage": "https://github.com/springload/draftail#readme",
-  "dependencies": {
-  },
+  "dependencies": {},
   "devDependencies": {
     "babel-core": "^6.18.2",
     "babel-loader": "^6.2.7",
@@ -34,6 +33,7 @@
     "babel-preset-react": "^6.16.0",
     "draft-js": "^0.9.1",
     "draftjs-utils": "^0.3.2",
+    "enzyme": "^2.6.0",
     "eslint": "^3.9.1",
     "eslint-config-airbnb": "^13.0.0",
     "eslint-plugin-import": "^2.2.0",
@@ -44,8 +44,9 @@
     "jest": "^17.0.0",
     "node-sass": "^3.11.2",
     "progress-bar-webpack-plugin": "^1.9.0",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "react": "^15.4.1",
+    "react-addons-test-utils": "^15.4.1",
+    "react-dom": "^15.4.1",
     "webpack": "^1.13.3",
     "webpack-dev-middleware": "^1.8.4",
     "webpack-hot-middleware": "^2.13.2"
