@@ -35,16 +35,28 @@ const blockRenderMap = DefaultDraftBlockRenderMap.merge(Map({
     },
 }));
 
+const defaultOptions = {
+    modelPickerOptions: [],
+    imageFormats: [],
+    mediaControls: [],
+    dialogControls: [],
+    sources: [],
+    decorators: [],
+    BLOCK_TYPES: [],
+    INLINE_STYLES: [],
+};
+
 /**
  * The top-level component for the Draft.js editor.
  */
 class DraftailEditor extends Component {
     constructor(props) {
         super(props);
-        const { options } = props;
+        const { rawContentState, options } = props;
+        const decorators = new CompositeDecorator(options.decorators);
 
         this.state = {
-            editorState: conversion.createEditorState(props.value, new CompositeDecorator(options.decorators)),
+            editorState: conversion.createEditorState(rawContentState, decorators),
         };
 
         let updateTimeout;
@@ -54,7 +66,7 @@ class DraftailEditor extends Component {
                 if (updateTimeout) {
                     global.clearTimeout(updateTimeout);
                 }
-                updateTimeout = global.setTimeout(this.saveRawState, STATE_SAVE_INTERVAL);
+                updateTimeout = global.setTimeout(this.saveState, STATE_SAVE_INTERVAL);
             });
         };
 
@@ -72,7 +84,7 @@ class DraftailEditor extends Component {
         this.renderControls = this.renderControls.bind(this);
 
         this.onRequestDialog = this.onRequestDialog.bind(this);
-        this.saveRawState = this.saveRawState.bind(this);
+        this.saveState = this.saveState.bind(this);
         this.setSelectionToCurrentEntity = this.setSelectionToCurrentEntity.bind(this);
         this.updateState = this.updateState.bind(this);
         this.onDialogComplete = this.onDialogComplete.bind(this);
@@ -111,10 +123,11 @@ class DraftailEditor extends Component {
       // }
     }
 
-    saveRawState() {
+    saveState() {
+        const { onSave } = this.props;
         const { editorState } = this.state;
 
-        this.inputRef.value = conversion.serialiseEditorState(editorState);
+        onSave(conversion.serialiseEditorState(editorState));
     }
 
     // Sets a selection to encompass the containing entity.
@@ -162,8 +175,6 @@ class DraftailEditor extends Component {
 
     updateState(state) {
         this.onChange(state);
-      // not sure we need this.
-      // setTimeout(() => this.editorRef.focus(), 0);
         return true;
     }
 
@@ -172,43 +183,39 @@ class DraftailEditor extends Component {
         const newState = RichUtils.onTab(event, editorState, MAX_LIST_NESTING);
         const isAtStart = newState.isSelectionAtStartOfContent();
         const isAtEnd = newState.isSelectionAtEndOfContent();
+        let ret;
 
         if (isAtStart && event.shiftKey) {
-            return;
+            ret = undefined;
+        } else if (isAtEnd) {
+            ret = undefined;
+        } else if (newState) {
+            ret = this.updateState(newState);
         }
 
-        if (isAtEnd) {
-            return;
-        }
-
-        if (newState) {
-            return this.updateState(newState);
-        }
+        return ret;
     }
 
     handleKeyCommand(command) {
         const { editorState } = this.state;
         const newState = RichUtils.handleKeyCommand(editorState, command);
+        let ret = false;
 
         if (newState) {
-            return this.updateState(newState);
+            ret = this.updateState(newState);
         }
 
-        return false;
+        return ret;
     }
 
-    toggleBlockType(blockType, e) {
+    toggleBlockType(blockType) {
         const { editorState } = this.state;
-
-        e.preventDefault();
 
         this.onChange(RichUtils.toggleBlockType(editorState, blockType));
     }
 
-    toggleInlineStyle(inlineStyle, e) {
+    toggleInlineStyle(inlineStyle) {
         const { editorState } = this.state;
-
-        e.preventDefault();
 
         this.onChange(RichUtils.toggleInlineStyle(editorState, inlineStyle));
     }
@@ -490,14 +497,13 @@ class DraftailEditor extends Component {
     }
 
     render() {
-        const { name } = this.props;
         const { editorState, readOnly } = this.state;
 
         return (
             <div
                 ref={(ref) => { this.wrapperRef = ref; }}
                 className="json-text"
-                onBlur={this.saveRawState}
+                onBlur={this.saveState}
                 onClick={this.handleFocus}
                 onMouseUp={this.onMouseUp}
                 onKeyUp={this.onKeyUp}
@@ -519,20 +525,14 @@ class DraftailEditor extends Component {
                 {this.renderTooltip()}
 
                 {this.renderDialog()}
-
-                <input
-                    ref={(ref) => { this.inputRef = ref; }}
-                    type="hidden"
-                    name={name}
-                />
             </div>
         );
     }
 }
 
 DraftailEditor.propTypes = {
-    name: React.PropTypes.string.isRequired,
-    value: React.PropTypes.string.isRequired,
+    rawContentState: React.PropTypes.object,
+    onSave: React.PropTypes.func,
     options: React.PropTypes.shape({
         modelPickerOptions: React.PropTypes.arrayOf(React.PropTypes.shape({
             label: React.PropTypes.string.isRequired,
@@ -547,7 +547,41 @@ DraftailEditor.propTypes = {
             label: React.PropTypes.string.isRequired,
             value: React.PropTypes.string.isRequired,
         })).isRequired,
-    }).isRequired,
+        mediaControls: React.PropTypes.arrayOf(React.PropTypes.shape({
+            entity: React.PropTypes.string.isRequired,
+            label: React.PropTypes.string.isRequired,
+            icon: React.PropTypes.string.isRequired,
+        })).isRequired,
+        dialogControls: React.PropTypes.arrayOf(React.PropTypes.shape({
+            entity: React.PropTypes.string.isRequired,
+            label: React.PropTypes.string.isRequired,
+            icon: React.PropTypes.string.isRequired,
+        })).isRequired,
+        sources: React.PropTypes.arrayOf(React.PropTypes.shape({
+            entity: React.PropTypes.string.isRequired,
+            control: React.PropTypes.func.isRequired,
+        })).isRequired,
+        decorators: React.PropTypes.arrayOf(React.PropTypes.shape({
+            strategy: React.PropTypes.func.isRequired,
+            component: React.PropTypes.func.isRequired,
+        })).isRequired,
+        BLOCK_TYPES: React.PropTypes.arrayOf(React.PropTypes.shape({
+            style: React.PropTypes.string.isRequired,
+            label: React.PropTypes.string.isRequired,
+            icon: React.PropTypes.string,
+        })).isRequired,
+        INLINE_STYLES: React.PropTypes.arrayOf(React.PropTypes.shape({
+            style: React.PropTypes.string.isRequired,
+            label: React.PropTypes.string.isRequired,
+            icon: React.PropTypes.string.isRequired,
+        })).isRequired,
+    }),
+};
+
+DraftailEditor.defaultProps = {
+    rawContentState: {},
+    onSave: () => {},
+    options: defaultOptions,
 };
 
 export default DraftailEditor;
