@@ -1,13 +1,12 @@
 import { EditorState, convertFromRaw, convertToRaw } from 'draft-js';
 
 export default {
-    // Serialised RawDraftContentState => EditorState.
-    createEditorState(serialisedState, decorators) {
-        const rawState = JSON.parse(serialisedState);
+    // RawDraftContentState + decorators => EditorState.
+    createEditorState(rawContentState, decorators) {
         let editorState;
 
-        if (rawState && Object.keys(rawState).length !== 0) {
-            const contentState = convertFromRaw(rawState);
+        if (rawContentState && Object.keys(rawContentState).length !== 0) {
+            const contentState = convertFromRaw(rawContentState);
             editorState = EditorState.createWithContent(contentState, decorators);
         } else {
             editorState = EditorState.createEmpty(decorators);
@@ -16,17 +15,17 @@ export default {
         return editorState;
     },
 
-    // EditorState => Serialised RawDraftContentState.
+    // EditorState => RawDraftContentState.
     serialiseEditorState(editorState) {
         const contentState = editorState.getCurrentContent();
-        const rawContent = convertToRaw(contentState);
+        const rawContentState = convertToRaw(contentState);
 
-        const isEmpty = rawContent.blocks.every((block) => {
+        const isEmpty = rawContentState.blocks.every((block) => {
             return block.text.trim().length === 0
                 && block.entityRanges.length === 0
                 && block.inlineStyleRanges.length === 0;
         });
 
-        return JSON.stringify(isEmpty ? {} : rawContent);
+        return isEmpty ? {} : rawContentState;
     },
 };
