@@ -4,8 +4,8 @@ import conversion from '../api/conversion';
 const emptyDecorator = new CompositeDecorator([]);
 
 const stubs = {
-    emptyContent: JSON.stringify({}),
-    realContent: JSON.stringify({
+    emptyContent: {},
+    realContent: {
         entityMap: {},
         blocks: [
             {
@@ -27,7 +27,7 @@ const stubs = {
                 data: {},
             },
         ],
-    }),
+    },
 };
 
 describe('conversion', () => {
