@@ -1,13 +1,18 @@
 import React from 'react';
 
+const onMouseDown = (onClick, e) => {
+    e.preventDefault();
+
+    onClick();
+};
+
 const Button = ({ icon, label, active, onClick }) => (
-    <a
-        href="#"
+    <button
         className={`RichEditor-styleButton${active ? ' RichEditor-activeButton' : ''}${icon ? ` icon icon-${icon}` : ''}`}
-        onClick={onClick}
+        onMouseDown={onMouseDown.bind(null, onClick)}
     >
         {label}
-    </a>
+    </button>
 );
 
 Button.propTypes = {
@@ -15,6 +20,11 @@ Button.propTypes = {
     onClick: React.PropTypes.func.isRequired,
     icon: React.PropTypes.string,
     active: React.PropTypes.bool,
+};
+
+Button.defaultProps = {
+    icon: '',
+    active: false,
 };
 
 export default Button;
