@@ -22,7 +22,7 @@ export default {
 
     getEntityRange: DraftUtils.getEntityRange.bind(DraftUtils),
 
-    getEntityData: (entityKey) => {
+    getEntityData(entityKey) {
         const entity = Entity.get(entityKey);
         return entity.getData();
     },
@@ -30,7 +30,7 @@ export default {
     /**
      * Returns the type of the block the current selection starts in.
      */
-    getSelectedBlockType: (editorState) => {
+    getSelectedBlockType(editorState) {
         const selectionState = editorState.getSelection();
         const contentState = editorState.getCurrentContent();
         const startKey = selectionState.getStartKey();
@@ -42,7 +42,7 @@ export default {
     /**
      * Creates a selection for the entirety of an entity that can be partially selected.
      */
-    getSelectedEntitySelection: (editorState) => {
+    getSelectedEntitySelection(editorState) {
         const selectionState = editorState.getSelection();
         const contentState = editorState.getCurrentContent();
         const entityKey = this.getSelectionEntity(editorState);
@@ -61,14 +61,14 @@ export default {
         });
     },
 
-    hasCurrentInlineStyle: (editorState, style) => {
+    hasCurrentInlineStyle(editorState, style) {
         const currentStyle = editorState.getCurrentInlineStyle();
 
         return currentStyle.has(style);
     },
 
     // TODO Document.
-    insertBlock: (editorState, entityKey, character, blockType) => {
+    insertBlock(editorState, entityKey, character, blockType) {
         const contentState = editorState.getCurrentContent();
         const selectionState = editorState.getSelection();
 
@@ -110,7 +110,7 @@ export default {
     },
 
     // TODO Document.
-    createEntity: (editorState, entityType, entityData, entityText, entityMutability = 'IMMUTABLE') => {
+    createEntity(editorState, entityType, entityData, entityText, entityMutability = 'IMMUTABLE') {
         const entityKey = Entity.create(entityType, entityMutability, entityData);
 
         const contentState = editorState.getCurrentContent();
